FoCo2: Coherent Forecast Combination for Linearly Constrained Multiple Time Series

Methods and tools designed to improve the forecast accuracy for a linearly constrained multiple time series, while fulfilling the linear/aggregation relationships linking the components (Girolimetto and Di Fonzo, 2024 doi:10.48550/arXiv.2412.03429 ). 'FoCo2' offers multi-task forecast combination and reconciliation approaches leveraging input from multiple forecasting models or experts and ensuring that the resulting forecasts satisfy specified linear constraints. In addition, linear inequality constraints (e.g., non-negativity of the forecasts) can be imposed, if needed.

See also

Author

Maintainer: Daniele Girolimetto daniele.girolimetto@unipd.it (ORCID) [funder, copyright holder]

Authors:

Tommaso Di Fonzo (ORCID) [funder]